Move package ecosystem code into a new pkgecosystem package.
